requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
CounterFit==0.1.4.dev9
counterfit-connection==0.1.0.dev5
counterfit-shims-grove==0.1.4.dev5
counterfit-shims-seeed-python-dht==0.1.0.dev1
dnspython==2.6.1
eventlet==0.33.3
Flask==2.1.2
Flask-SocketIO==5.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
paho-mqtt==2.1.0
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.32.3
ruff==0.5.1
simple-websocket==1.0.0
six==1.16.0
urllib3==2.2.2
Werkzeug==2.3.6
wsproto==1.2.0
zipp==3.19.2
